+++ .github/workflows/tests-download.yaml
@@ -35,4 +35,4 @@ jobs:
         run: poetry install --no-interaction --with dev
 
       - name: Run Download Tests
-        run: poetry run pytest -ra -v -m "download"
+        run: poetry run pytest -ra -v -x -m "download"
