+++ src/auto_archiver/archivers/__init__.py
@@ -1,12 +1,9 @@
 from .archiver import Archiver
 from .telethon_archiver import TelethonArchiver
-from .twitter_archiver import TwitterArchiver
 from .twitter_api_archiver import TwitterApiArchiver
 from .instagram_archiver import InstagramArchiver
 from .instagram_tbot_archiver import InstagramTbotArchiver
-from .tiktok_archiver import TiktokArchiver
 from .telegram_archiver import TelegramArchiver
 from .vk_archiver import VkArchiver
-from .youtubedl_archiver import YoutubeDLArchiver
+from .base_archiver.base_archiver import BaseArchiver as YoutubeDLArchiver
 from .instagram_api_archiver import InstagramAPIArchiver
-from .bluesky_archiver import BlueskyArchiver
