+++ src/auto_archiver/archivers/base_archiver/__init__.py
@@ -1,0 +1,1 @@
+from .base_archiver import BaseArchiver
